backend/requirements.txt
aioredis==2.0.1
amqp==5.1.1
aniso8601==9.0.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
camel-converter==2.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.3
django-cors-headers==3.13.0
django-countries==7.4.2
django-crequest==2018.5.11
django-filter==22.1
django-graphql-jwt==0.3.4
django-import-export==3.0.1
django-seed==0.3.1
django-versatileimagefield==2.2
et-xmlfile==1.1.0
Faker==15.3.2
graphene==3.1.1
graphene-django==3.0.0
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
h11==0.14.0
hiredis==2.0.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
MarkupPy==1.14
meilisearch==0.22.2
msgpack==1.0.4
numpy==1.23.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
promise==2.3
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-magic==0.4.27
pytimeparse==1.1.8
pytz==2022.6
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
Rx==3.2.0
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tablib==3.2.1
text-unidecode==1.3
toposort==1.7
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.4
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.5.2
python-decouple==3.6